feat: add to performance drag and drop
