+++ src/DraggableSchedule.tsx
@@ -1,0 +1,74 @@
+import {
+  Box,
+  Popover,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverArrow,
+  PopoverCloseButton,
+  PopoverBody,
+  Button,
+  Text,
+} from "@chakra-ui/react";
+import { CSS } from "@dnd-kit/utilities";
+import { useDraggable } from "@dnd-kit/core";
+import { ComponentProps, memo } from "react";
+import { DAY_LABELS, CellSize } from "./constants";
+import { Lecture } from "./types/type";
+
+export const DraggableSchedule = memo(
+  ({
+    id,
+    bg,
+    day,
+    range,
+    room,
+    lecture,
+    onDeleteButtonClick,
+  }: { id: string; lecture: Lecture; day: string; range: number[]; room?: string } & ComponentProps<typeof Box> & {
+      onDeleteButtonClick: () => void;
+    }) => {
+    const { attributes, setNodeRef, listeners, transform } = useDraggable({ id });
+    const leftIndex = DAY_LABELS.indexOf(day as (typeof DAY_LABELS)[number]);
+    const topIndex = range[0] - 1;
+    const size = range.length;
+
+    return (
+      <Popover>
+        <PopoverTrigger>
+          <Box
+            position="absolute"
+            left={`${120 + CellSize.WIDTH * leftIndex + 1}px`}
+            top={`${40 + (topIndex * CellSize.HEIGHT + 1)}px`}
+            width={CellSize.WIDTH - 1 + "px"}
+            height={CellSize.HEIGHT * size - 1 + "px"}
+            bg={bg}
+            p={1}
+            boxSizing="border-box"
+            cursor="pointer"
+            ref={setNodeRef}
+            transform={CSS.Translate.toString(transform)}
+            {...listeners}
+            {...attributes}
+          >
+            <Text fontSize="sm" fontWeight="bold">
+              {lecture.title}
+            </Text>
+            <Text fontSize="xs">{room}</Text>
+          </Box>
+        </PopoverTrigger>
+        <PopoverContent onClick={(event) => event.stopPropagation()}>
+          <PopoverArrow />
+          <PopoverCloseButton />
+          <PopoverBody>
+            <Text>강의를 삭제하시겠습니까?</Text>
+            <Button colorScheme="red" size="xs" onClick={onDeleteButtonClick}>
+              삭제
+            </Button>
+          </PopoverBody>
+        </PopoverContent>
+      </Popover>
+    );
+  },
+);
+
+export default DraggableSchedule;
