+++ src/ScheduleDndProvider.tsx
@@ -47,11 +47,11 @@ export default function ScheduleDndProvider({ children }: PropsWithChildren) {
     const moveDayIndex = Math.floor(x / 80);
     const moveTimeIndex = Math.floor(y / 30);
 
-    setSchedulesMap({
-      ...schedulesMap,
-      [tableId]: schedulesMap[tableId].map((targetSchedule, targetIndex) => {
+    setSchedulesMap(prev => ({
+      ...prev,
+      [tableId]: prev[tableId].map((targetSchedule, targetIndex) => {
         if (targetIndex !== Number(index)) {
-          return { ...targetSchedule };
+          return targetSchedule;
         }
         return {
           ...targetSchedule,
@@ -59,7 +59,7 @@ export default function ScheduleDndProvider({ children }: PropsWithChildren) {
           range: targetSchedule.range.map((time) => time + moveTimeIndex),
         };
       }),
-    });
+    }));
   };
 
   return (
