+++ src/ScheduleTables.tsx
@@ -3,6 +3,8 @@ import ScheduleTable from "./ScheduleTable.tsx";
 import { useScheduleContext } from "./ScheduleContext.tsx";
 import SearchDialog from "./SearchDialog.tsx";
 import { useState } from "react";
+import ScheduleDndProvider from "./ScheduleDndProvider.tsx";
+import { useAutoCallback } from "./hooks/useAutoCallback.ts";
 
 export const ScheduleTables = () => {
   const { schedulesMap, setSchedulesMap } = useScheduleContext();
@@ -28,6 +30,10 @@ export const ScheduleTables = () => {
     })
   };
 
+  const handleScheduleTimeClick = useAutoCallback((tableId: string, timeInfo: { day: string; time: number }) => {
+    setSearchInfo({ tableId, ...timeInfo });
+  });
+
   return (
     <>
       <Flex w="full" gap={6} p={6} flexWrap="wrap">
@@ -42,16 +48,18 @@ export const ScheduleTables = () => {
                         onClick={() => remove(tableId)}>삭제</Button>
               </ButtonGroup>
             </Flex>
-            <ScheduleTable
-              key={`schedule-table-${index}`}
-              schedules={schedules}
-              tableId={tableId}
-              onScheduleTimeClick={(timeInfo) => setSearchInfo({ tableId, ...timeInfo })}
-              onDeleteButtonClick={({ day, time }) => setSchedulesMap((prev) => ({
-                ...prev,
-                [tableId]: prev[tableId].filter(schedule => schedule.day !== day || !schedule.range.includes(time))
-              }))}
-            />
+            <ScheduleDndProvider>
+              <ScheduleTable
+                key={`schedule-table-${index}`}
+                schedules={schedules}
+                tableId={tableId}
+                onScheduleTimeClick={useAutoCallback((timeInfo) => handleScheduleTimeClick(tableId, timeInfo))}
+                onDeleteButtonClick={({ day, time }) => setSchedulesMap((prev) => ({
+                  ...prev,
+                  [tableId]: prev[tableId].filter(schedule => schedule.day !== day || !schedule.range.includes(time))
+                }))}
+              />
+            </ScheduleDndProvider>
           </Stack>
         ))}
       </Flex>
